The test found my intentional error and failed. Fixing the intentional error

+++ src/software_title/script.py
@@ -3,9 +3,9 @@ import click
 @click.command()
 @click.option("--a", type=str, required=True, help="A string")
 @click.option(
-    "--b", type=int, required=True, help="An integer"
+    "--b", type=float, required=True, help="An integer"
 )
-def func(a:str, b: int, c: float = 1.5):
+def func(a:str, b: float, c: float = 1.5):
     """_summary_
 
     Args:
